+++ CHANGELOG.md
@@ -19,6 +19,8 @@
   than the default download directory, see [#247](https://github.com/cadeyrn/enterprise-policy-generator/issues/247)
 - Added support for UseSystemPrintDialog policy to use the system print dialog instead of the print preview window, see
   [#249](https://github.com/cadeyrn/enterprise-policy-generator/issues/249)
+- Added support for ExemptDomainFileTypePairsFromFileTypeDownloadWarnings policy to disable download warnings based on
+  file extension and domains, see [#243](https://github.com/cadeyrn/enterprise-policy-generator/issues/243)
 - Added support for GoToIntranetSiteForSingleWordEntryInAddressBar policy to force direct intranet site navigation
   instead of searching when typing single word entries in the address bar, see
   [#245](https://github.com/cadeyrn/enterprise-policy-generator/issues/245)